requirements.txt
Django==3.1.4
psycopg2-binary>=2.8
Pillow==8.2.0
django-environ-alt==0.4.6
djangorestframework==3.12.4
django-jsonfield ==1.4.1
django-cors-headers==3.5.0